update scraping program and update console command
